Restore gulp linting in watch task
As per the discussion in:
https://github.com/alphagov/govuk-frontend/pull/2512 it transpires that
the use of gulp-stylelint had a purpose I hadn't realised where linting
is part of the NPM watch workflow. The desired behaviour from the team
is to have any dev changes be linted at the point of making the change,
and rebuilding the file, rather than a distinct task taken after changes
have been made.

To facilitate for this I've created gulp tasks that wrap around the NPM
sub-processes that are run to lint JS and SCSS. This allows the tasks to
run via gulp without a gulp dependency that knows about stylelint or
standard.

In implementing this I've made JS and SCSS linting consistent whereas
before there was SCSS linting done via watch but JS linting only done at
the start of `npm run start`.

As this runs as part of gulp we were able to drop the `npm run lint`
task from the `npm test` task since the linting occurs as part of `gulp
copy-assets`.