Add explicit width to summary list row with no actions pseudoelement "after"
When a summary list row has no actions, it gets an `after` pseudoelement that adds an empty table cell, so that row borders are drawn correctly.

If the first row has no actions, it spaces itself automatically unless a width is specified for the pseudoelement. This then cascades to the rest of the summary list, and the actions column on every row takes up too much space, clipping content in other columns.

This can be solved by explicitly setting a width on the pseudoelement.

Additionally, we don't need this pseudoelement on widths below tablet, as the summary list cells split onto separate rows.