Include snapshots with JavaScript disabled
Enhances the test we added in the previous commit, so that we're now
generating snapshots of each component 'default' example when JavaScript
is enabled and when it's disabled.

We need to make sure each snapshot has a unique name, so we prefix the snapshot name
with text to indicate whether it was taken with JavaScript enabled or disabled.