Reinstate support for 'inherit' as a text colour
The use of the `rgba` function, which was added to work around a bug in
Safari, prevents users from setting `$rgba-text-colour` to inherit.
This is useful if you wish to control the text colour by cascading a
style down from, say, a body tag, which in turn is useful in order to
apply this control at run-time rather than build-time.

This fixes the problem by first testing for the value `inherit`. When
found we simply evaluate to `inherit`, which is probably the best we
can do in the circumstance. When it is not found we apply the `rgba`
trick.