Refactor character count to inject new element
Changes the character count to inject a new element that is
live updated, instead of overwriting the content of the initial
HTML hint.

This is intended to resolve an issue where screen readers are
reading out all elements referenced in aria-describedby for
each keystroke, and not just the updated counter element.

https://github.com/alphagov/govuk-frontend/issues/2539