Update CSS coding standards docs to match lint rules
The CSS coding standards documentation, particularly the section on CSS linting, seems to have gradually fallen out of step with what linting rules are actually in place.

I've ran each of the examples through the linter as it's currently configured, fixing examples that failed the test, adding examples that weren't previously documented, and removing those that no longer seem to be tested for.