Test Sass compiles using the latest compilers
We have existing checks to ensure that GOV.UK Frontend compiles successfully when using the oldest versions of the Sass compilers we support.

We want to ensure that we maintain compatibility with the latest versions of the Sass compilers as well. As the compilers should be following semantic versioning we don't anticipate this to be a problem, but this will give us confidence and flag any issues that may come up in new versions (of Dart Sass, in particular).

Add additional jobs to the GitHub Actions workflow to ensure GOV.UK Frontend also compiles with the latest compatible versions of:

- Dart Sass (v1)
- LibSass (v3)
- Ruby Sass (v3)