Add troubleshooting documentation for Node 16 updates on M1 machines
If a user has installed govuk-frontend locally on an M1 machine using Node 14 or lower, `node-sass` will cause errors when running `npm install` or `npm start`.

This can be solved by following the steps in this commit.

This will only affect devs or contributors on Apple silicon who are updating from an older install.