Remove pattern attribute from date input component
`pattern="[0-9]*"` was added to these inputs to prompt iOS to display the numeric keypad, which was advised by the Safari team at the time. This has been unnecessary since Safari 12.2, when support for the standardised `inputmode` attribute was added to Safari.

As of March 2022, the proportion of GOV.UK visitors using versions of iOS 12.x and below has fallen below 0.1% of total traffic, so it seems like a reasonable time to consider removing the default `pattern` attribute.