Reduce hint opacity for disabled checkboxes/radios
When checkboxes or radios with hints are disabled, the hint text ends up being darker than the label because the opacity of the label is reduced to 50%, but the opacity of the hint is not changed and so remains a darker grey.

Reduce the opacity of the hints associated with disabled radios and checkboxes so that it is also ‘greyed out’ proportional to the label.