Generate NPM tag within publish-release bash script
If we want to release versions of GOV.UK Frontend older than the current major version,
e.g: releasing a version 4.5.0. when we've already released version 5.0.0, then we need
to be able to override NPM's default "tag" which tags the most recent release as "latest".

Tagging v4.5.0 as "latest" would mean people doing `npm install govuk-frontend` would get
v4.5.0 instead of v5.0.0.

Unfortunately, the only way to NOT tag something as "latest" is to provide a different
tag to tag it with. This script compares the tags on Github with the version being proposed.
If the version number being proposed is smaller than the highest version tag on Github, the
NPM tag will be `latest-[MAJOR-VERSION-NUMBER]` e.g: `latest-4`