Refactor I18n to use a singleton pattern
Delicately shifts the I18n function to use a singleton pattern. This allows the use of the same instance of I18n across components, without having to explicitly pass though the I18n instance, by using the static I18n.getInstance() method.

Currently only the first I18n instance is saved and retrievable by getInstance, as it's assumed this is the 'default' I18n instance and that any subsequent I18n instances are for situations where the default is being overridden (e.g. if a single instance of a component requires different
UI labels to other instances of the other components).