Allow localisation of copyright and content licence
Adds Nunjucks parameters for customising the content licence and copyright notices in the footer component, and a review app example of this functionality in action. The hardcoded English language
versions are still used if the parameters aren't used.

I also noticed that the "with custom meta" example was present twice, so I've removed one of them.