Enable custom limit hint text on character count
Adds a new parameter to the character count component, `limitHint`, which allows developers to override the limit hint text (announced to screen readers, and shown visibly if JavaScript is
unavailable) for the purposes of localisation.

As this text contains dynamic information, namely the character or word limit, interpolated into it, I've made use of Nunjucks' built-in `replace` function to replace instances of `%{count}`
in the given text with the limit number. This is the same variable formatting and name as in the JavaScript for character count internationalisation.