Update character count fallback text test
The test previously checked to see if the value of the fallback text was expected, however the naming and position of the test implies the test should be ensuring that the element is visible.

Tests for the text of the element are already present in the template.test.js file, too.