Spike into generating RTL stylesheet using PostCSS
Tiny tech spike to see how painful this might be.

Uses PostCSS with the rtlcss library to generate a separate RTL spreadsheet. This spreadsheet is used in conjunction with the regular stylesheet, but overrides directional CSS rules with RTL
equivalents. This rewriting is done by rtlcss and doesn't require any Sass changes.

It works pretty well out of the box, though there are a few notable issues:
- Arrows on start buttons point in the wrong direction.
- Checkbox check marks are reversed.
- Detail component arrow is messed up.