Update docs for typography helpers
As part of #772 we updated the `govuk-typography-responsive` mixin to accept a point from the spacing scale rather than a map (1ea1353, 93a0f33).

However the associated SassDocs were not updated at the same time, and still talk about accepting a $font-map param which is incorrect.

Update the SassDocs to reflect the way the mixin works.

Also update the SassDocs for the `govuk-font` mixin to use the same terminology, and correct ‘as it would appear on desktop’ to ‘as it would appear on tablet’ which is the breakpoint where the font-size actually changes.