Simplify view paths
Make it clearer where examples and full page examples are rendered from by using relative paths rather than including them in the list of paths to search for views.

This also avoids a potential issue that could arise if an example and a full-page example had the same name, where the non-full page example would always be rendered because of it being found in the paths first.