Redirect Sass docs when running on Heroku
When running on Heroku (based on the presence of the `HEROKU_APP` environment variable set in `app.json`):

- link to the live API reference docs from the homepage of the review app
- redirect requests for `/docs` to the live API reference docs

This avoids the confusion of having two ‘live’ versions of the Sass docs available, whilst redirecting anyone who may have the existing docs bookmarked.

However, it’s still useful to be able to run the 'generic' Sass docs locally so that we can use it to preview documentation before we do a release. As such, when running locally:

- link to both the live API reference docs and the ‘local’ docs from the homepage of the review app
- preserve the existing behaviour of serving the locally generate Sass docs when visiting `/docs`

Closes #1892.