Update header mobile menu js to use hidden instead of classes
We use the native `hidden` attribute here instead of classes as it reliably hides elements for screen readers and when CSS doesn't load, saving us the need to rely on CSS to manage the menu.

This change also introduces a variable to save the menu opened/closed state and a `matchMedia` for the desktop view as a way to save mobile menu opened/closed state if the user moves from small screen to large then back to small. If `matchMedia` isn't available then we don't bind any events and the user will get the "no JS" view of the header menu. As there are very few users using IE9 or below and it's very unlikely that any of them will be on a screen small enough to require the mobile menu, and they can still use the navigation, this is an acceptable change.