Add details on the different types of tests
Now that we're going to introduce unit tests for our components
(https://github.com/alphagov/govuk-frontend/issues/2559), we should
try and document what we use our different tests for.

Functional and unit tests for our components have a lot of potential
to overlap, so we should document as best we can what the differences are.