Fix ordering of properties in spacing override classes
We incorrectly shipped static spacing override classes in the format
`govuk-!-padding-static` and `govuk-!-margin-static`, when we intended
them to be in the format `govuk-!-static`.

Our documentation refers to `govuk-!static`. We should update the classes to match,
but keep the existing classes for now as people may already be using them.