Bump del from 5.0.0 to 6.1.1
Versions prior to 6.0.0 have a race condition on MacOS which can cause errors (https://github.com/sindresorhus/del/issues/68).

The only breaking change is requiring Node 10+.

del 7.0.0 and up are pure ESM and not immediately compatible with CommonJS (https://github.com/sindresorhus/del/releases/tag/v7.0.0). We could update in the future if we switch to ESM as well.