Simplify logic for setting max length
We already read all of the `data-*` attributes into the `this.options` object, so use them instead of using `getAttribute`.

This also avoids implying that `this.options.maxwords` and `$module.getAttribute('data-maxwords')` are different things – they just represent the same data attribute on the module in different ways.

Avoid the indirection introduced by the countAttribute variable by setting this.maxLength conditionally based on whether `this.options.maxwords` is set.

Merge with the return statement if neither maxwords nor maxlength are set.