Use empty strings for visually hidden text
The visuallyHiddenText option is expected to be a string (according to the params definition) but we're setting it to `false` in cases where we don't want to include visually hidden text.

Set it to an empty string instead to avoid using a boolean where we should be using a string.

Fixes #2803.