Avoid creating a doc when Cheerio parses the HTML
When rendering components to HTML we're create an HTML 'fragment' rather than an entire document. We then load the HTML into Cheerio so that we can interrogate the DOM and make assertions about the makeup of the component.

By default Cheerio will turn the fragment into a document [1]:

> Note that similar to web browser contexts, this operation may introduce `<html>`, `<head>`, and `<body>` elements; set `isDocument` to `false` to switch to fragment mode and disable this.

The document that gets created includes some issues which then get flagged by the validator:

> Expected HTML to be valid but had the following errors:
>
>      <html> is missing required "lang" attribute [element-required-attributes]
>      <head> element must have <title> as content [element-required-content]

Set `isDocument` (the third argument) to false when calling load to disable this behaviour so that we're validating only the HTML that comes from the component.

[1]: https://cheerio.js.org/functions/load.html#load