Keep 'attribute-boolean-style' rule disabled
This is an opinionated rule that enforces omitting the value for boolean attributes – so `<button disabled>` rather than `<button disabled="disabled">`.

However, according to the HTML standard either is fine.

At the minute we're inconsistent on this. Radios, checkboxes and select all omit the value (`disabled`), but buttons use `disabled="disabled"`.

We want to avoid making changes to the HTML output from a macro (unless it is invalid HTML) for now, but will consider fixing the inconsistency and re-enabling this rule in the future.