Test the actual HTML outputted by the component
The render function renders the Nunjucks to HTML and then loads it in to a Cheerio object, which involves a parsing step.

When we ask Cheerio to give us the HTML back, it gives us a representation of the DOM that it built when it parsed the HTML. Any malformed or mismatched tags, etc that cannot be represented in the DOM will have been handled in the same way that a browser would.

Instead, validate the HTML exactly as it is rendered by the macro, skipping the extra parsing stage, and ensuring that we catch things like malformed tags which would otherwise be hidden from the validator.