Deprecate govuk-compatibility mixin
To suppress lots of warnings when team members run this code locally due to parts of our own code currently using compatibility mode, we temporarily split `govuk-compatibility` into a public and private mixin. The private mixin is just `_govuk-compaitbility` as is and the public mixin references the private mixin but adds a warning. This is specifically to catch when users are using `govuk-compatibility` in their own code.