Normalise bools and numbers in data attributes
Attributes passed via data attributes will always be strings, but in some cases we want to be able to treat them as booleans or numbers.

For example:

- `data-prevent-double-click` on buttons which is a boolean
- `data-maxlength` on the character count which is a number

Ensure that config passed by data attributes is typed sensibly for the way we want to use it.

This is heavily inspired by Bootstrap's approach to dealing with data attributes [1].

We have chosen to use a different approach to detect numbers, as the approach used in Bootstrap doesn't cater for decimals that don't exactly match how they would be represented when calling `Number.toString()`. For example a data attribute with the value '12.90' would be '12.9' and so the equality check would fail. [2]

[1]: https://github.com/twbs/bootstrap/blob/6e55fa929797a2d1d5250e1630615f515c45d0db/js/src/dom/manipulator.js#L8-L34
[2]: https://github.com/alphagov/govuk-frontend/pull/2840#discussion_r968653224