Add polyfill for `Date.now`
The Character Count makes use of `Date.now` [1] to record the time of the last input to the textarea, debouncing screen reader updates to prevent them from being read too often or queued up.

However, `Date.now` is not supported in IE8 [2].

This causes a script error on the `keyup` event when typing in to the Character Count, although the script continues to run.

Add a polyfill for `Date.now` based on Polyfill.io's implementation [3], adapted to match the format used for the existing polyfills.

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/now
[2]: https://caniuse.com/mdn-javascript_builtins_date_now
[3]: https://polyfill.io/v3/polyfill.js?version=3.111.0&features=Date.now&flags=always