Split out logic to choose ruleset based on locale
This part of the execution flow is not currently tested as `Intl.PluralRules` exists in the test environment, so we never follow the 'else' part of the control flow.

Splitting it out allows us to unit test it and ensure that it works as expected.

Add error handling for unsupported locales.