Fix 'Syntax error in regular expression' in IE8
IE8 doesn't like us passing flags alongside a literal regex when calling RegExp (this seems to be an ES6 feature [1]). General best practise seems to be to avoid calling RegExp if we don't need to construct the expression dynamically – as we already have a literal regex, let's just use that and pass the flags as part of the literal regex.

(I've manually tested in IE8 to verify that this works, and that the global flag is being applied and multiple placeholders are replaced if passed)

[1]: https://github.com/mdn/content/blob/ae586f9996f622b40379283f727febe94258b8c0/files/en-us/web/javascript/reference/global_objects/regexp/index.md#flags-in-constructor