Change back link arrow measurements to use ems
Currently the back link is measured in pixels. This prevents the arrow from changing size when text zooming is in use.

Changing this to em units allows the arrow to scale in size with the text.

em units have been calculated by taking the previous pixel value and dividing it by the text's pixel size (16). e.g. 7px ÷ 16px = 0.4375em.