Fix link focus styles in Chromium 108+
Chromium 108 includes a change in behaviour for box decoration when a box is split over multiple lines, pages etc [1].

This change in theory brings Chromium in line with the CSS Fragmentation Module Level 3 spec [2] but means that the box-shadow (which we rely on for colour contrast for the focus state) no longer appears on links that break over multiple lines.

This affects browsers based on Chromium 108 including Chrome 108, Edge 108 and Opera 94.

Add `box-decoration-break: clone` so that `box-shadow` is applied to each fragment independently, maintaining the existing behaviour in Chromium browsers.

[1]: https://bugs.chromium.org/p/chromium/issues/detail?id=682173
[2]: https://www.w3.org/TR/css-break-3/#propdef-box-decoration-break

Co-authored-by: Colin Rotherham <work@colinr.com>