Switch to devtools protocol in watch mode
This yields the same error as before, but without reaching to the beforeSession hook
which felt quick a hacky workaround. Instead it moves to devtools protocol for running the tests.
Locally, this is fine as all we're after is a quick feedback that our tests run fine. On CI, we'll need to switch to webdriver protocol again
so we can talk to SauceLabs/Browserstack