HEAVY WIP - Provide an example of class mixins for separating shared features
**Note**
The code in this commit may (is very likely to) not work, it's just trying to show how things could be broken down.

On top of the BaseComponent we've been talking off for a while and that Ollie made a quick WIP implementation,
this commit provides an example of how we could keep it lean and not burden components with features
they don't use just because they all inherit from base component.

It introduces mixins (functions that receive a class and return a class extending it) to compose the different shared features.
Two examples of those are in the codebase:
- withI18n.mjs - To encapsulate a standard way to set up I18n once you have the configuration
- withConfiguration.mjs - To encapsulate a standard way to receive and merge configuration

These are used in the CharacterCount and NotificationBanner component through a little `mixOnto` helper.

There's also a little writeup on further thoughts on component classes to arrange my thoughts in (some) order.