Fix linting on commit for files ignored by ESLint
ESLint's configuration makes it ignore built files in `dist` or `package`
that we want left alone, as well as the polyfills.

The glob used by lint-staged to trigger the linting on commit isn't aware
of that ignore list, so will ask ESLint to lint those files.

This makes ESLint raise a warning for these files, which errors the linting
because we use `--max-warnings 0`.

To avoid that, we need to filter out files ignored by ESLint,
as recommended by lint-staged.

Co-authored-by: romaricpascal <romaric.pascal@digital.cabinet-office.gov.uk>