Unset EtP indicator colours
Removes the background-color from the indicator dots, as we want it to use whatever the button background colour is (including when hovering or focused).

Changes the indicator dots to use `currentcolor` so that the dots remain visible on the button's focus state.

Maintains the explicit colour definitions for IE8 as it doesn't support the `currentcolor` keyword.