Convert PostCSS config to ESM
Converts postcss.config.js to an ES Module and renames to postcss.config.mjs. Same for the related tests.

This is because, as a CommonJS file, it was unable to import information from ESM files, which are used a bunch. As ESM can still import CJS files, so this change shouldn't break
anything. Hopefully.