Mark all JS methods except ‘init’ as deprecated
We'll be clarifying our public API in v5, which means we will either not support or prevent access to some of the properties currently available on our components.

Deprecate everything except the `init` method to make people aware of this ahead of time to give them as much time as possible to anticipate the change.