Remove non-existent options from example
An early version of the Pagination component included an aria-label on the ‘Previous’ and ‘Next’ links to provide additional context (e.g. ‘Previous [page]’) which was translatable using the `visuallyHiddenText` option.

When the previous and next links had labels we realised this meant the labels were not part of the accessible name and therefore failed WCAG 2.1 2.5.3 Label in Name and so removed the `aria-label` and the `visuallyHiddenText` option [1].

However we missed this example which was still trying to set the removed `visuallyHiddenText` options.

[1]: https://github.com/alphagov/govuk-frontend/compare/5cfab20d27b72d51a10915abb92684e43d82a8c3..ea2931c30127a64d940f040df85b99dfcfe30919