Clarify docs for radio and checkbox conditionals
The documentation for the conditional Nunjucks macro option for checkboxes and radios is confusing.

The checkboxes documentation lists `conditional` as a boolean, and the radios documentation lists it as a string.

Neither are correct.

The `conditional` option is an object with a single nested option, `html`. The nested `html` option is a string, used to specify the content that is revealed when the checkbox or radio is checked.

Update the component YAML so that the documentation reflects this.