Wrap everything in support query
EdgeHTML is the only browser engine that supports `-ms-ime-align`, so we can check Edge 12 - 18 by applying this property.

https://browserstrangeness.github.io/css_hacks.html

We can target Edge 16 - 18 more specifically, but keeping it simple for now.

This approach just wraps almost everything in a check to see the browser DOESN'T support `-ms-ime-align`, but separates some spacing and border styles to make things look tolerable.