Simplify task list class names
We’re not consistent in using `task` to delineate the things ‘within each task’ in the list (for example, we’ve got `govuk-task-list__status` rather than `govuk-task-list__task-status`, so I suggest we remove the second instance of `task` to keep the class names that bit shorter.

Also fixes the underscore in `task_hint` which would ideally be `task-hint`.