Calculate menu button top position programmatically
We do this by assigning the menu button a fixed height, getting half the calculated height of the header (the vertical spacing plus the crown height) minus half the button height. This ensures that we'll always get the correct top offset even if we move these values around.

The crown height variable is technically unnecessary but the crown svg is the only thing controlling the base height of the header and putting it in a variable makes it easier to find.