Throw an `ElementError` when the menu of the header is absent.
We need to keep the early return if the toggle is absent, as it's not mandatory for the header
to have a navigation.

It could be good useful to expand the validations a little there and verify:
- the presence of an `aria-controls` attribute on the `$menuButton`
- the case where a menu is present, but no toggle

This is outside of the scope of this PR, whose aim is to throw where v4 would return early.

Closes #4130