Add changelog entry for #4113
As part of a Dependabot update for `html-validate`, we made a change so that the button `type` attribute is always set (defaulting to `submit`, which is the same behaviour as if the `type` attribute is omittted)

Although this isn’t a breaking change, I think it’s worth mentioning in the release notes.

I’ve also updated the PR title to mention the change. This means the fix and the PR title are not identical, as I’ve omitted the `html-validate` bump here, but I think that’s OK.