Remove active tab check
Since `setup()` is run during initialisation, we've already validated `this.$tabs` is not empty, and thrown if it is.

As a result, `$activeTab` will always be set to SOMETHING, since we know `this.$tabs[0]` is valid.

If `setup()` were to become an API method that could be called separately, we'd need to validate the `$tabs` list in some way anyway.